Create sagemaker_train_vit.py
We add a training function with the correct structure for training on AWS Sage Master

Some additional points to note:

You may want to adjust the __getitem__ method in AestheticDataset to suit how you've set up your labels. This implementation assumes that the label is in the filename of each image.

The data channel arguments (--train, --val, --test) expect the paths to the data directories. When you start a training job, SageMaker will automatically download your data from S3 to these directories.

The --model_dir argument is where the trained model (model.pth) will be saved. SageMaker will automatically upload this file back to your S3 bucket once training is complete.
You may need to adjust the hyperparameters and their default values in the argument parser to suit your needs.

Remember to handle the case where your labels are not just part of the image filename. You may have to change how labels are read in the __getitem__ method.

Ensure that your data is correctly labeled and formatted before training.